+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.3
+
+- Fixed an issue where the inner scroll view exceeded the boundaries of the header ([issue#3](https://github.com/idootop/nested_scroll_view_plus/issues/4)).
+- Corrected behavior of NestedScrollView to ensure it follows the defined physics.
+- Synchronized NestedScrollView source code with Flutter version [Flutter 3.19.0-5.0.pre](https://github.com/flutter/flutter/commit/e5f62cc5a029469f46464a6930075731ce42a94d).
+
+
 ## 1.0.2
 
 - Fixed the performLayout of RenderSliverOverlapAbsorberInner